+++ packages/cli/src/cli.ts
@@ -4,9 +4,9 @@ import * as fs from 'fs';
 import * as minimatch from 'minimatch';
 import { gzip } from 'node-gzip';
 import { program } from '@caporal/core';
-import { Logger, NodeIO } from '@gltf-transform/core';
+import { Logger, NodeIO, PropertyType } from '@gltf-transform/core';
 import { ALL_EXTENSIONS } from '@gltf-transform/extensions';
-import { AOOptions, CenterOptions, DedupOptions, InstanceOptions, PartitionOptions, PruneOptions, ResampleOptions, SequenceOptions, UnweldOptions, WeldOptions, ao, center, dedup, instance, metalRough, partition, prune, resample, sequence, unweld, weld } from '@gltf-transform/lib';
+import { AOOptions, CenterOptions, InstanceOptions, PartitionOptions, PruneOptions, ResampleOptions, SequenceOptions, UnweldOptions, WeldOptions, ao, center, dedup, instance, metalRough, partition, prune, resample, sequence, unweld, weld } from '@gltf-transform/lib';
 import { inspect } from './inspect';
 import { DracoCLIOptions, ETC1S_DEFAULTS, Filter, Mode, UASTC_DEFAULTS, draco, merge, toktx, unlit } from './transforms';
 import { Session, formatBytes } from './util';
@@ -198,9 +198,14 @@ compression and instancing, to be more effective.
 		validator: program.BOOLEAN,
 		default: true,
 	})
-	.action(({args, options, logger}) =>
-		Session.create(io, logger, args.input, args.output)
-			.transform(dedup(options as unknown as DedupOptions))
+	.action(({args, options, logger}) => {
+		const propertyTypes: string[] = [];
+		if (options.accessors) propertyTypes.push(PropertyType.ACCESSOR);
+		if (options.textures) propertyTypes.push(PropertyType.TEXTURE);
+		return Session.create(io, logger, args.input, args.output)
+			.transform(dedup({propertyTypes}));
+	}
+
 	);
 
 // PRUNE